lab2/requirements.txt
requests
pymongo
pyyaml
beautifulsoup4
lxml